+++ R/validate_model_opt.R
@@ -25,7 +25,7 @@ validate_model_opt <- function(model_opt_input) {
     mixing = FALSE,
     draws_type = "scrambled_sobol",
     R = 10,
-    fixed = NULL,
+    fixed = c(""),
     param = NULL,
     rpar = NULL
   )
@@ -42,6 +42,23 @@ validate_model_opt <- function(model_opt_input) {
   if (is.null(model_opt$N) || is.null(model_opt$S) || is.null(model_opt$J)) {
     cat(red$bold(symbol$cross), "  model_opt().\n")
     stop("You must specify the number of respondents 'N', maximum number of choice occasions 'S' and the maximum number of alternatives 'J'. These can be functions of your data.")
+  }
+  
+  # Check the fixed parameter options ----
+  # The implementation of fixed parameters is strongly influenced by the 'apollo package' and these tests are also implemented there.
+  if (length(model_opt$fixed) > 0 && !is.character(model_opt$fixed)) {
+    cat(red$bold(symbol$cross), "  model_opt().\n")
+    stop("'fixed' is a character vector with the names of the parameters to keep fixed at their starting values during estimation.")
+  }
+  
+  if (length(unique(model_opt$fixed)) < length(model_opt$fixed)) {
+    cat(red$bold(symbol$cross), "  model_opt().\n")
+    stop("'fixed' contains duplicate elements. Please check that all parameters are unique.")
+  }
+  
+  if (!all(model_opt$fixed %in% names(model_opt$param))) {
+    cat(red$bold(symbol$cross), "  model_opt().\n")
+    stop("Some parameters included in 'fixed' is not specified in the list of parameters. Please make sure that the parameters to fix are parameters in your model.")
   }
   
   # Check mixing options ----