+++ NEWS.md
@@ -1,4 +1,5 @@
 ## cmdlR v0.0.1
+* Can specify paramters to stay fixed at the starting values using model_opt$fixed
 * Added convenient function store() governed by the list save_opt() to save results and model objects
 * Added working LC example
 * Added working MIXL example